Added function to calculate effective bandwidth
The function calculates the effective bandwidth by parsing the spw
selection string. The number of output spws are then calculated given
the effective bandwidth.

The prior method was to calculate the number of output spws based on the
entire bandwidth in the MS ignoring the spw selection. This would result
in the number of output spws spanning well beyond the physical
limitations of the telescope.